+++ agents/stock_price_agent/data_manager.py
@@ -1,43 +1,31 @@
 """
-Stock Price Agent Data Manager
-키움 API 응답 데이터 저장, 필터링, 관리 (새로운 로직)
+Simplified data manager for ChatClovaX agent
+Saves chart data to JSON files and returns pandas DataFrames
+Includes upgrade suggestions when data is insufficient for requested date range
 """
 
 import os
 import json
-import shutil
+import pandas as pd
+import pandas_ta as ta
 from datetime import datetime
-from typing import Dict, List, Any, Optional
 from pathlib import Path
+from typing import Dict, Any, Optional
 
 
 class StockDataManager:
-    """키움 API 데이터 관리자 (새로운 로직)"""
+    """Simplified data manager for stock chart data with upgrade suggestions for insufficient data"""
     
     def __init__(self, base_path: str = None):
-        """
-        데이터 매니저 초기화
-        
-        Args:
-            base_path: 데이터 저장 기본 경로
-        """
         if base_path is None:
-            # agents/stock_price_agent/data/ 경로 설정
             current_dir = Path(__file__).parent
             base_path = current_dir / "data"
         
         self.data_dir = Path(base_path)
-        self.raw_dir = self.data_dir / "raw"          # 원본 데이터
-        self.filtered_dir = self.data_dir / "filtered" # 필터링된 데이터
+        self.raw_dir = self.data_dir / "raw"
+        self.filtered_dir = self.data_dir / "filtered"
         
-        # 차트 유형별 API 함수 매핑 초기화
-        self._init_chart_configs()
-        
-        # 폴더 초기화
-        self.initialize_directories()
-    
-    def _init_chart_configs(self):
-        """차트 유형별 API 함수 매핑 초기화 (틱 차트 제거)"""
+        # Chart type configurations
         self.chart_configs = {
             "minute": {
                 "api_function": "ka10080",
@@ -65,40 +53,269 @@ class StockDataManager:
                 "date_field": "dt"
             }
         }
+        
+        # Technical indicator configurations for different chart types
+        self.indicator_configs = self._setup_indicator_configs()
+        
+        self._init_directories()
     
-    def initialize_directories(self):
-        """데이터 디렉토리 초기화"""
-        try:
-            # 기존 data 폴더가 있으면 삭제
-            if self.data_dir.exists():
-                print(f"🗑️  기존 데이터 폴더 삭제: {self.data_dir}")
-                shutil.rmtree(self.data_dir, ignore_errors=True)
-            
-            # 새 폴더 생성
-            self.raw_dir.mkdir(parents=True, exist_ok=True)
-            self.filtered_dir.mkdir(parents=True, exist_ok=True)
-            
-            print(f"📁 데이터 폴더 초기화 완료:")
-            print(f"   • Raw: {self.raw_dir}")
-            print(f"   • Filtered: {self.filtered_dir}")
-            
-        except Exception as e:
-            print(f"❌ 데이터 폴더 초기화 실패: {e}")
+    def _setup_indicator_configs(self):
+        """Setup technical indicator parameters for different chart types and timeframes"""
+        return {
+            "minute_1_3_5": {
+                "sma_periods": [9, 20, 50],
+                "ema_periods": [9, 20, 50], 
+                "macd": {"fast": 6, "slow": 13, "signal": 5},
+                "rsi": 9,
+                "stoch": {"k": 9, "d": 3},
+                "bollinger": {"n": 20, "std": 2},
+                "atr": 9,
+                "cmf": 20
+            },
+            "minute_10_15": {
+                "sma_periods": [20, 50, 100],
+                "ema_periods": [20, 50, 100],
+                "macd": {"fast": 12, "slow": 26, "signal": 9},
+                "rsi": 14,
+                "stoch": {"k": 14, "d": 3},
+                "bollinger": {"n": 20, "std": 2},
+                "atr": 14,
+                "cmf": 20
+            },
+            "minute_30_45_60": {
+                "sma_periods": [20, 50, 200],
+                "ema_periods": [20, 50, 200],
+                "macd": {"fast": 12, "slow": 26, "signal": 9},
+                "rsi": 14,
+                "stoch": {"k": 14, "d": 3},
+                "bollinger": {"n": 20, "std": 2},
+                "atr": 14,
+                "cmf": 20
+            },
+            "day": {
+                "sma_periods": [20, 50, 200],
+                "ema_periods": [20, 50, 200],
+                "macd": {"fast": 12, "slow": 26, "signal": 9},
+                "rsi": 14,
+                "stoch": {"k": 14, "d": 3},
+                "bollinger": {"n": 20, "std": 2},
+                "atr": 14,
+                "cmf": 20
+            },
+            "week": {
+                "sma_periods": [10, 20, 50],
+                "ema_periods": [10, 20, 50],
+                "macd": {"fast": 6, "slow": 13, "signal": 5},
+                "rsi": 9,
+                "stoch": {"k": 9, "d": 3},
+                "bollinger": {"n": 10, "std": 2},
+                "atr": 10,
+                "cmf": 10
+            },
+            "month": {
+                "sma_periods": [6, 12, 24],
+                "ema_periods": [6, 12, 24],
+                "macd": {"fast": 3, "slow": 6, "signal": 3},
+                "rsi": 6,
+                "stoch": {"k": 6, "d": 3},
+                "bollinger": {"n": 6, "std": 2},
+                "atr": 6,
+                "cmf": 6
+            },
+            "year": {
+                "sma_periods": [3, 5, 10],
+                "ema_periods": [3, 5, 10],
+                "macd": {"fast": 2, "slow": 5, "signal": 2},
+                "rsi": 6,
+                "stoch": {"k": 6, "d": 3},
+                "bollinger": {"n": 6, "std": 2},
+                "atr": 6,
+                "cmf": 6
+            }
+        }
     
-    def save_raw_data(self, raw_data: Dict[str, Any], stock_code: str, chart_type: str, base_date: str = None) -> str:
+    def _init_directories(self):
+        """Initialize data directories"""
+        self.raw_dir.mkdir(parents=True, exist_ok=True)
+        self.filtered_dir.mkdir(parents=True, exist_ok=True)
+    
+    def process_chart_data(self, raw_data: Dict[str, Any], stock_code: str, 
+                          chart_type: str, base_date: str = None,
+                          expected_start_date: str = None, 
+                          expected_end_date: str = None,
+                          minute_scope: str = None) -> Dict[str, Any]:
         """
-        원본 데이터 저장 (새로운 파일명 규칙)
+        Process chart data from Kiwoom API with upgrade suggestions when data is insufficient
         
         Args:
-            raw_data: 키움 API 원본 응답
-            stock_code: 종목코드
-            chart_type: 차트 유형
-            base_date: 기준일자 (YYYYMMDD)
+            raw_data: Raw API response
+            stock_code: Stock symbol (e.g., "005930")
+            chart_type: Chart type (minute, day, week, month, year)
+            base_date: Base date for the request
+            expected_start_date: Expected start date (YYYYMMDD)
+            expected_end_date: Expected end date (YYYYMMDD)
+            minute_scope: Minute scope for minute charts
             
         Returns:
-            str: 저장된 파일 경로
+            Dict: Processing result with data or upgrade suggestions for insufficient data
         """
-        # 새로운 파일명 규칙: {요청시간}_{api-id}_{주식종목}_{base_date}.json
+        # 1. Save raw data with proper naming convention
+        raw_filepath = self._save_raw_data(raw_data, stock_code, chart_type, base_date)
+        
+        # 2. Convert ALL raw data to DataFrame with date format conversion
+        df = self._extract_chart_dataframe(raw_data, chart_type)
+        
+        if df.empty:
+            return {
+                "status": "no_data",
+                "message": "원본 데이터가 비어있습니다.",
+                "data": None
+            }
+        
+        # 3. Add technical indicators to FULL dataset
+        df = self._add_technical_indicators(df, chart_type, minute_scope)
+        
+        # 4. Apply date filtering if requested
+        if expected_start_date and expected_end_date:
+            # Check oldest date in dataframe first
+            oldest_date = self._find_oldest_date_in_dataframe(df, chart_type)
+            
+            if oldest_date and oldest_date > expected_start_date:
+                # Data doesn't go back far enough - suggest upgrade
+                upgrade_info = self._get_chart_upgrade_suggestion(chart_type)
+                
+                return {
+                    "status": "upgrade_required",
+                    "message": f"데이터 부족: 가장 오래된 데이터({oldest_date}) > 요청 시작일({expected_start_date})",
+                    "upgrade_suggestion": upgrade_info,
+                    "original_start_date": expected_start_date,
+                    "original_end_date": expected_end_date,
+                    "data": None
+                }
+            
+            # Filter dataframe by date range
+            df_filtered = self._filter_dataframe_by_date_range(
+                df, chart_type, expected_start_date, expected_end_date
+            )
+            
+            if df_filtered.empty:
+                return {
+                    "status": "no_data",
+                    "message": "지정된 날짜 범위에 데이터가 없습니다.",
+                    "data": None
+                }
+            
+            # Check if filtered records are sufficient (minimum 10 records)
+            if len(df_filtered) < 10:
+                downgrade_info = self._get_chart_downgrade_suggestion(chart_type, minute_scope)
+                
+                return {
+                    "status": "downgrade_required",
+                    "message": f"레코드 부족: {len(df_filtered)}개 < 10개 최소 요구량",
+                    "downgrade_suggestion": downgrade_info,
+                    "original_start_date": expected_start_date,
+                    "original_end_date": expected_end_date,
+                    "data": None
+                }
+            
+            # Use filtered dataframe
+            df = df_filtered
+        
+        # Apply chart-specific date format after filtering
+        df = self._convert_date_format_for_chart_type(df, chart_type)
+        
+        # 5. Save processed data to CSV
+        filtered_filepath = self._save_filtered_data_csv(
+            df, stock_code, chart_type, base_date, expected_start_date, expected_end_date
+        )
+        
+        return {
+            "status": "success",
+            "message": f"데이터 처리 완료: {len(df)}개 레코드",
+            "data": df
+        }
+    
+    def format_tool_response(self, result: Dict[str, Any], stock_code: str, chart_type: str) -> str:
+        """
+        Unified response formatting for all tools
+        Handles upgrade suggestions for insufficient data and DataFrame table formatting
+        
+        Args:
+            result: Result from process_chart_data()
+            stock_code: Stock symbol
+            chart_type: Chart type description (e.g., "daily", "5-minute")
+            
+        Returns:
+            str: Formatted response string
+        """
+        status = result.get("status", "unknown")
+        
+        if status == "upgrade_required":
+            upgrade_info = result.get("upgrade_suggestion", {})
+            next_type = upgrade_info.get("next_type")
+            description = upgrade_info.get("description", "")
+            original_start = result.get("original_start_date")
+            original_end = result.get("original_end_date")
+            
+            response = f"{result.get('message', '')} \n\n"
+            response += f"업그레이드 제안: {description}\n"
+            if next_type:
+                if original_start and original_end:
+                    response += f"권장 툴: get_{next_type}_chart(stock_code='{stock_code}', expected_start_date='{original_start}', expected_end_date='{original_end}')\n"
+                else:
+                    response += f"권장 툴: get_{next_type}_chart(stock_code='{stock_code}')\n"
+            
+            return response
+        
+        elif status == "downgrade_required":
+            downgrade_info = result.get("downgrade_suggestion", {})
+            next_type = downgrade_info.get("next_type")
+            next_scope = downgrade_info.get("next_scope")
+            description = downgrade_info.get("description", "")
+            original_start = result.get("original_start_date")
+            original_end = result.get("original_end_date")
+            
+            response = f"{result.get('message', '')} \n\n"
+            response += f"다운그레이드 제안: {description}\n"
+            if next_type == "minute" and next_scope:
+                if original_start and original_end:
+                    response += f"권장 툴: get_minute_chart(stock_code='{stock_code}', minute_scope='{next_scope}', expected_start_date='{original_start}', expected_end_date='{original_end}')\n"
+                else:
+                    response += f"권장 툴: get_minute_chart(stock_code='{stock_code}', minute_scope='{next_scope}')\n"
+            elif next_type and next_type != "minute":
+                if original_start and original_end:
+                    response += f"권장 툴: get_{next_type}_chart(stock_code='{stock_code}', expected_start_date='{original_start}', expected_end_date='{original_end}')\n"
+                else:
+                    response += f"권장 툴: get_{next_type}_chart(stock_code='{stock_code}')\n"
+            
+            return response
+        
+        elif status == "success":
+            df = result.get("data")
+            if df is not None and not df.empty:
+                response = f"✅ **{chart_type} 차트 데이터** ({stock_code}):\n\n"
+                response += self._format_dataframe_table(df)
+                return response
+            else:
+                return f"No {chart_type} chart data available for {stock_code}"
+        
+        elif status == "no_data":
+            return f"❌ {result.get('message', 'No data available')}"
+        
+        else:
+            return f"❌ Unknown status: {status}"
+    
+    def _format_dataframe_table(self, df: pd.DataFrame) -> str:
+        """Format DataFrame as complete table without summary info"""
+        if df.empty:
+            return "No data available"
+        
+        # Return complete DataFrame as table string
+        return df.to_string(index=False, max_cols=None, max_rows=None)
+    
+    def _save_raw_data(self, raw_data: Dict[str, Any], stock_code: str, 
+                      chart_type: str, base_date: str = None) -> str:
+        """Save raw data with proper naming convention: {요청시간}_{api-id}_{주식종목}_{base_date}.json"""
         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
         api_id = self.chart_configs.get(chart_type, {}).get("api_function", "unknown")
         base_date_str = base_date if base_date else "nodate"
@@ -107,63 +324,129 @@ class StockDataManager:
         filepath = self.raw_dir / filename
         
         try:
-            # 원본 JSON을 한 글자도 다르지 않게 그대로 저장
             with open(filepath, 'w', encoding='utf-8') as f:
                 json.dump(raw_data, f, ensure_ascii=False, indent=2)
-            
-            print(f"💾 원본 데이터 저장: {filepath}")
+            print(f"Raw data saved: {filepath}")
             return str(filepath)
-            
         except Exception as e:
-            print(f"❌ 원본 데이터 저장 실패: {e}")
+            print(f"❌ Failed to save raw data: {e}")
             return ""
     
-    def _find_oldest_date_in_raw_data(self, raw_data: Dict[str, Any], chart_type: str) -> Optional[str]:
+    def _find_oldest_date_in_dataframe(self, df: pd.DataFrame, chart_type: str) -> Optional[str]:
         """
-        Raw 데이터에서 가장 오래된 날짜 찾기
+        Find the oldest date in DataFrame
         
         Args:
-            raw_data: 키움 API 원본 응답
-            chart_type: 차트 유형
+            df: DataFrame with date column (YYYYMMDD string format)
+            chart_type: Chart type
             
         Returns:
-            str: 가장 오래된 날짜 (YYYYMMDD) 또는 None
+            str: Oldest date in YYYYMMDD format or None
         """
-        if chart_type not in self.chart_configs:
+        if df.empty or 'date' not in df.columns:
             return None
         
-        config = self.chart_configs[chart_type]
-        data_key = config["data_key"]
-        date_field = config["date_field"]
-        
-        if data_key not in raw_data or not raw_data[data_key]:
+        try:
+            # All dates are now strings (YYYYMMDD format)
+            oldest_date_str = df['date'].min()
+            if pd.isna(oldest_date_str) or not isinstance(oldest_date_str, str):
+                return None
+            
+            # For all chart types, return the date string as-is or extract YYYYMMDD
+            if chart_type in ["minute", "day"]:
+                # For minute/day data, date should be YYYYMMDD or YYYYMMDDHHMISS
+                if len(oldest_date_str) >= 8:
+                    return oldest_date_str[:8]  # Extract YYYYMMDD
+                return oldest_date_str
+            
+            elif chart_type == "week":
+                # For weekly data, extract YYYYMMDD from YYYYMMWeekN format if already converted
+                # or return YYYYMMDD if not yet converted
+                if "Week" in oldest_date_str and len(oldest_date_str) >= 6:
+                    return oldest_date_str[:6] + "01"  # Return first day of month
+                elif len(oldest_date_str) >= 8:
+                    return oldest_date_str[:8]  # YYYYMMDD
+                return oldest_date_str
+            
+            elif chart_type in ["month", "year"]:
+                # For monthly/yearly data, should be YYYYMMDD format before conversion
+                if len(oldest_date_str) >= 8:
+                    return oldest_date_str[:8]  # YYYYMMDD
+                return oldest_date_str
+            
+            return oldest_date_str
+            
+        except Exception as e:
+            print(f"❌ Error finding oldest date in dataframe: {e}")
             return None
+    
+    def _filter_dataframe_by_date_range(self, df: pd.DataFrame, chart_type: str,
+                                       expected_start_date: str, expected_end_date: str) -> pd.DataFrame:
+        """
+        Filter DataFrame by date range
         
-        oldest_date = None
+        Args:
+            df: DataFrame with date column (YYYYMMDD string format)
+            chart_type: Chart type
+            expected_start_date: Start date (YYYYMMDD)
+            expected_end_date: End date (YYYYMMDD)
+            
+        Returns:
+            pd.DataFrame: Filtered DataFrame
+        """
+        if df.empty or 'date' not in df.columns:
+            return pd.DataFrame()
         
-        for record in raw_data[data_key]:
-            if isinstance(record, dict) and date_field in record:
-                record_date = record[date_field]
-                if record_date:
-                    # 시간 정보가 있으면 제거 (YYYYMMDD만 추출)
-                    if len(record_date) > 8:
-                        record_date = record_date[:8]
-                    
-                    if oldest_date is None or record_date < oldest_date:
-                        oldest_date = record_date
-        
-        return oldest_date
+        try:
+            # All dates are now strings (YYYYMMDD format), so we can use string comparison
+            def extract_date_for_comparison(date_str, chart_type):
+                """Extract YYYYMMDD string for comparison"""
+                if pd.isna(date_str) or not isinstance(date_str, str):
+                    return None
+                
+                if chart_type in ["minute", "day"]:
+                    # For minute/day data, extract YYYYMMDD
+                    if len(date_str) >= 8:
+                        return date_str[:8]  # YYYYMMDD
+                    return date_str
+                
+                elif chart_type == "week":
+                    # For weekly data, could be YYYYMMWeekN or YYYYMMDD
+                    if "Week" in date_str and len(date_str) >= 6:
+                        # Extract YYYYMM and use first day of month
+                        return date_str[:6] + "01"  # YYYYMM + 01
+                    elif len(date_str) >= 8:
+                        return date_str[:8]  # YYYYMMDD
+                    return date_str
+                
+                elif chart_type in ["month", "year"]:
+                    # For monthly/yearly data, should be YYYYMMDD before conversion
+                    if len(date_str) >= 8:
+                        return date_str[:8]  # YYYYMMDD
+                    return date_str
+                
+                return date_str
+            
+            # Extract comparable dates
+            df_temp = df.copy()
+            df_temp['comparable_date'] = df_temp['date'].apply(
+                lambda x: extract_date_for_comparison(x, chart_type)
+            )
+            
+            # Filter using string comparison (YYYYMMDD strings can be compared lexicographically)
+            mask = (df_temp['comparable_date'] >= expected_start_date) & (df_temp['comparable_date'] <= expected_end_date)
+            mask = mask & df_temp['comparable_date'].notna()  # Remove NaN values
+            
+            filtered_df = df[mask].copy().reset_index(drop=True)
+            print(f"📅 DataFrame 날짜 필터링 완료: {len(filtered_df)}개 레코드")
+            return filtered_df
+            
+        except Exception as e:
+            print(f"❌ DataFrame 날짜 필터링 오류: {e}")
+            return pd.DataFrame()
     
     def _get_chart_upgrade_suggestion(self, current_chart_type: str) -> Dict[str, str]:
-        """
-        차트 간격 업그레이드 제안 (틱 차트 제거)
-        
-        Args:
-            current_chart_type: 현재 차트 유형
-            
-        Returns:
-            Dict: 업그레이드 제안 정보
-        """
+        """Get chart upgrade suggestion"""
         upgrade_map = {
             "minute": {"next_type": "day", "description": "분봉 → 일봉으로 차트 유형 변경"},
             "day": {"next_type": "week", "description": "일봉 → 주봉으로 차트 유형 변경"},
@@ -177,80 +460,184 @@ class StockDataManager:
             "description": "업그레이드 옵션이 없습니다."
         })
     
-    def _filter_data_by_date_range(self, raw_data: Dict[str, Any], chart_type: str, expected_start_date: str, expected_end_date: str) -> List[Dict[str, Any]]:
+    def _get_chart_downgrade_suggestion(self, current_chart_type: str, minute_scope: str = None) -> Dict[str, str]:
         """
-        날짜 범위에 맞는 데이터만 필터링하여 핵심 필드만 추출
+        Get chart downgrade suggestion for insufficient records
         
         Args:
-            raw_data: 키움 API 원본 응답
-            chart_type: 차트 유형
-            expected_start_date: 시작일 (YYYYMMDD)
-            expected_end_date: 종료일 (YYYYMMDD)
+            current_chart_type: Current chart type
+            minute_scope: Current minute scope for minute charts
             
         Returns:
-            List: 필터링된 차트 데이터
+            Dict: Downgrade suggestion info
         """
-        if chart_type not in self.chart_configs:
-            return []
-        
-        config = self.chart_configs[chart_type]
-        data_key = config["data_key"]
-        date_field = config["date_field"]
-        
-        if data_key not in raw_data or not raw_data[data_key]:
-            return []
-        
-        filtered_records = []
-        
-        try:
-            filter_start = datetime.strptime(expected_start_date, "%Y%m%d")
-            filter_end = datetime.strptime(expected_end_date, "%Y%m%d")
+        if current_chart_type == "minute" and minute_scope:
+            # 분봉 다운그레이드: 60→45→30→15→10→5→3→1
+            minute_val = int(minute_scope)
+            downgrade_map = {
+                60: {"next_scope": "45", "description": "60분봉 → 45분봉으로 차트 간격 축소"},
+                45: {"next_scope": "30", "description": "45분봉 → 30분봉으로 차트 간격 축소"},
+                30: {"next_scope": "15", "description": "30분봉 → 15분봉으로 차트 간격 축소"},
+                15: {"next_scope": "10", "description": "15분봉 → 10분봉으로 차트 간격 축소"},
+                10: {"next_scope": "5", "description": "10분봉 → 5분봉으로 차트 간격 축소"},
+                5: {"next_scope": "3", "description": "5분봉 → 3분봉으로 차트 간격 축소"},
+                3: {"next_scope": "1", "description": "3분봉 → 1분봉으로 차트 간격 축소"},
+                1: {"next_scope": None, "description": "1분봉이 최소 간격입니다. 기간을 늘리거나 분석 방법을 변경하세요."}
+            }
             
-            for record in raw_data[data_key]:
-                if isinstance(record, dict) and date_field in record:
-                    record_date_str = record[date_field]
-                    if record_date_str:
-                        # 시간 정보가 있으면 제거 (YYYYMMDD만 추출)
-                        if len(record_date_str) > 8:
-                            record_date_str = record_date_str[:8]
-                        
-                        try:
-                            record_date = datetime.strptime(record_date_str, "%Y%m%d")
-                            
-                            # 날짜 범위 체크
-                            if filter_start <= record_date <= filter_end:
-                                # 핵심 필드만 추출
-                                filtered_record = {
-                                    "cur_prc": record.get("cur_prc", ""),
-                                    "trde_qty": record.get("trde_qty", ""),
-                                    "trde_prica": record.get("trde_prica", ""),
-                                    "dt": record.get("dt", record.get("cntr_tm", "")),
-                                    "open_pric": record.get("open_pric", ""),
-                                    "high_pric": record.get("high_pric", ""),
-                                    "low_pric": record.get("low_pric", "")
-                                }
-                                filtered_records.append(filtered_record)
-                        except ValueError:
-                            continue
+            suggestion = downgrade_map.get(minute_val, {"next_scope": None, "description": "다운그레이드 옵션이 없습니다."})
+            if suggestion["next_scope"]:
+                suggestion["next_type"] = "minute"
+            else:
+                suggestion["next_type"] = None
             
-            print(f"📅 날짜 범위 필터링 완료: {len(filtered_records)}개 레코드")
-            return filtered_records
+            return suggestion
+        
+        else:
+            # 일반 차트 다운그레이드: 년→월→주→일→분(60분)
+            downgrade_map = {
+                "year": {"next_type": "month", "description": "년봉 → 월봉으로 차트 유형 변경"},
+                "month": {"next_type": "week", "description": "월봉 → 주봉으로 차트 유형 변경"},
+                "week": {"next_type": "day", "description": "주봉 → 일봉으로 차트 유형 변경"},
+                "day": {"next_type": "minute", "next_scope": "60", "description": "일봉 → 60분봉으로 차트 유형 변경"}
+            }
             
-        except ValueError as e:
-            print(f"❌ 날짜 파싱 오류: {e}")
-            return []
+            return downgrade_map.get(current_chart_type, {
+                "next_type": None, 
+                "description": "다운그레이드 옵션이 없습니다."
+            })
     
-    def _convert_date_format_for_chart_type(self, filtered_records: List[Dict[str, Any]], chart_type: str) -> List[Dict[str, Any]]:
+    def _extract_chart_dataframe(self, raw_data: Dict[str, Any], 
+                                chart_type: str) -> pd.DataFrame:
+        """Extract chart data and convert to DataFrame (for non-filtered data)"""
+        
+        # Map chart types to data keys and date fields
+        data_mapping = {
+            "minute": {"key": "stk_min_pole_chart_qry", "date_field": "cntr_tm"},
+            "day": {"key": "stk_dt_pole_chart_qry", "date_field": "dt"},
+            "week": {"key": "stk_stk_pole_chart_qry", "date_field": "dt"},
+            "month": {"key": "stk_mth_pole_chart_qry", "date_field": "dt"},
+            "year": {"key": "stk_yr_pole_chart_qry", "date_field": "dt"}
+        }
+        
+        if chart_type not in data_mapping:
+            return pd.DataFrame()
+        
+        mapping = data_mapping[chart_type]
+        data_key = mapping["key"]
+        date_field = mapping["date_field"]
+        
+        # Extract chart records
+        chart_records = raw_data.get(data_key, [])
+        if not chart_records:
+            return pd.DataFrame()
+        
+        # Convert to DataFrame (without date format conversion yet)
+        df = pd.DataFrame(chart_records)
+        
+        # Standardize column names
+        column_mapping = {
+            date_field: 'date',
+            'cur_prc': 'close',
+            'open_pric': 'open', 
+            'high_pric': 'high',
+            'low_pric': 'low',
+            'trde_qty': 'volume',
+            'trde_prica': 'amount'
+        }
+        
+        # Rename columns that exist
+        existing_columns = {k: v for k, v in column_mapping.items() if k in df.columns}
+        df = df.rename(columns=existing_columns)
+        
+        # Convert numeric columns
+        numeric_columns = ['close', 'open', 'high', 'low', 'volume', 'amount']
+        for col in numeric_columns:
+            if col in df.columns:
+                df[col] = pd.to_numeric(df[col], errors='coerce')
+        
+        # Keep date column as string (YYYYMMDD format)
+        if 'date' in df.columns:
+            df['date'] = df['date'].astype(str)
+        
+        # Select and reorder columns
+        standard_columns = ['date', 'open', 'high', 'low', 'close', 'volume', 'amount']
+        available_columns = [col for col in standard_columns if col in df.columns]
+        df = df[available_columns].copy()
+        
+        # Sort by date
+        if 'date' in df.columns:
+            df = df.sort_values('date').reset_index(drop=True)
+        
+        return df
+    
+    def _convert_date_format_for_chart_type(self, data, chart_type: str):
         """
         차트 유형에 맞게 날짜 형식을 변환합니다
         
         Args:
-            filtered_records: 필터링된 차트 데이터
+            data: 차트 데이터 (list of records 또는 pandas DataFrame)
             chart_type: 차트 유형
             
         Returns:
-            List: 날짜 형식이 변환된 차트 데이터
+            List or DataFrame: 날짜 형식이 변환된 차트 데이터 (입력 타입과 동일)
         """
+        # DataFrame인 경우
+        if isinstance(data, pd.DataFrame):
+            if data.empty or 'date' not in data.columns:
+                return data
+            
+            df_copy = data.copy()
+            
+            if chart_type == "minute":
+                # For minute data, convert to datetime format
+                df_copy['date'] = pd.to_datetime(df_copy['date'], format='%Y%m%d%H%M%S', errors='coerce')
+                
+            elif chart_type == "day":
+                # For daily data, convert to date format
+                df_copy['date'] = pd.to_datetime(df_copy['date'], format='%Y%m%d', errors='coerce').dt.date
+            
+            elif chart_type == "week":
+                # DataFrame의 date 컬럼을 YYYYMMWeekN 형식으로 변환
+                def date_to_week_format(date_str):
+                    if pd.isna(date_str) or not isinstance(date_str, str):
+                        return str(date_str)
+                    if len(date_str) >= 8:  # YYYYMMDD 형식
+                        year_month = date_str[:6]  # YYYYMM
+                        day = date_str[6:8]
+                        week_num = (int(day) - 1) // 7 + 1
+                        return f"{year_month}Week{week_num}"
+                    return str(date_str)
+                
+                df_copy['date'] = df_copy['date'].apply(date_to_week_format)
+                
+            elif chart_type == "month":
+                # DataFrame의 date 컬럼을 YYYYMM 형식으로 변환
+                def date_to_month_format(date_str):
+                    if pd.isna(date_str) or not isinstance(date_str, str):
+                        return str(date_str)
+                    if len(date_str) >= 6:  # YYYYMMDD 형식
+                        return date_str[:6]  # YYYYMM
+                    return str(date_str)
+                
+                df_copy['date'] = df_copy['date'].apply(date_to_month_format)
+                
+            elif chart_type == "year":
+                # DataFrame의 date 컬럼을 YYYY 형식으로 변환
+                def date_to_year_format(date_str):
+                    if pd.isna(date_str) or not isinstance(date_str, str):
+                        return str(date_str)
+                    if len(date_str) >= 4:  # YYYYMMDD 형식
+                        return date_str[:4]  # YYYY
+                    return str(date_str)
+                
+                df_copy['date'] = df_copy['date'].apply(date_to_year_format)
+            
+            print(f"📅 DataFrame 날짜 형식 변환 완료: {chart_type} 차트용")
+            return df_copy
+        
+        # List인 경우 (기존 로직 유지)
+        filtered_records = data
         if not filtered_records or chart_type not in ["minute", "day", "week", "month", "year"]:
             return filtered_records
         
@@ -310,195 +697,190 @@ class StockDataManager:
         print(f"📅 날짜 형식 변환 완료: {chart_type} 차트용 {len(converted_records)}개 레코드")
         return converted_records
     
-    def save_filtered_data(self, filtered_records: List[Dict[str, Any]], stock_code: str, chart_type: str, base_date: str = None) -> str:
+    def _get_indicator_config_key(self, chart_type: str, minute_scope: str = None) -> str:
+        """Get indicator configuration key based on chart type and minute scope"""
+        if chart_type == "minute" and minute_scope:
+            minute_val = int(minute_scope)
+            if minute_val in [1, 3, 5]:
+                return "minute_1_3_5"
+            elif minute_val in [10, 15]:
+                return "minute_10_15"
+            elif minute_val in [30, 45, 60]:
+                return "minute_30_45_60"
+        
+        return chart_type  # day, week, month, year
+    
+    def _add_technical_indicators(self, df: pd.DataFrame, chart_type: str, minute_scope: str = None) -> pd.DataFrame:
         """
-        필터링된 데이터 저장
+        Add technical indicators to DataFrame based on chart type and timeframe
         
         Args:
-            filtered_records: 필터링된 차트 데이터
-            stock_code: 종목코드
-            chart_type: 차트 유형
-            base_date: 기준일자
+            df: DataFrame with OHLCV data
+            chart_type: Chart type (minute, day, week, month, year)
+            minute_scope: Minute scope for minute charts (1, 3, 5, 10, 15, 30, 45, 60)
             
         Returns:
-            str: 저장된 파일 경로
+            pd.DataFrame: DataFrame with technical indicators added
         """
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        base_date_str = base_date if base_date else "nodate"
+        if df.empty or len(df) < 10:  # Need minimum data for indicators
+            print("⚠️  Insufficient data for technical indicators")
+            return df
         
-        filename = f"{timestamp}_{chart_type}_{stock_code}_{base_date_str}_filtered.json"
-        filepath = self.filtered_dir / filename
+        # Get indicator configuration
+        config_key = self._get_indicator_config_key(chart_type, minute_scope)
+        config = self.indicator_configs.get(config_key, self.indicator_configs["day"])
+        
+        # Create a copy to avoid modifying original
+        df_with_indicators = df.copy()
         
         try:
-            with open(filepath, 'w', encoding='utf-8') as f:
-                json.dump(filtered_records, f, ensure_ascii=False, indent=2)
+            # Ensure required columns exist and are numeric
+            required_cols = ['open', 'high', 'low', 'close', 'volume']
+            for col in required_cols:
+                if col not in df_with_indicators.columns:
+                    print(f"⚠️  Missing {col} column for technical indicators")
+                    return df
+                df_with_indicators[col] = pd.to_numeric(df_with_indicators[col], errors='coerce')
             
-            print(f"✅ 필터링된 데이터 저장: {filepath}")
-            return str(filepath)
+            # Remove rows with NaN values in OHLCV columns
+            df_with_indicators = df_with_indicators.dropna(subset=required_cols)
+            
+            if len(df_with_indicators) < 10:
+                print("⚠️  Insufficient valid data after cleaning")
+                return df
+            
+            # Calculate SMA (Simple Moving Averages)
+            for period in config["sma_periods"]:
+                if len(df_with_indicators) >= period:
+                    df_with_indicators[f'sma_{period}'] = ta.sma(df_with_indicators['close'], length=period)
+            
+            # Calculate EMA (Exponential Moving Averages)
+            for period in config["ema_periods"]:
+                if len(df_with_indicators) >= period:
+                    df_with_indicators[f'ema_{period}'] = ta.ema(df_with_indicators['close'], length=period)
+            
+            # Calculate MACD
+            macd_params = config["macd"]
+            if len(df_with_indicators) >= macd_params["slow"]:
+                macd_result = ta.macd(
+                    df_with_indicators['close'], 
+                    fast=macd_params["fast"], 
+                    slow=macd_params["slow"], 
+                    signal=macd_params["signal"]
+                )
+                if macd_result is not None and not macd_result.empty:
+                    df_with_indicators = pd.concat([df_with_indicators, macd_result], axis=1)
+            
+            # Calculate RSI
+            if len(df_with_indicators) >= config["rsi"]:
+                df_with_indicators['rsi'] = ta.rsi(df_with_indicators['close'], length=config["rsi"])
+            
+            # Calculate Stochastic
+            stoch_params = config["stoch"]
+            if len(df_with_indicators) >= stoch_params["k"]:
+                stoch_result = ta.stoch(
+                    df_with_indicators['high'], 
+                    df_with_indicators['low'], 
+                    df_with_indicators['close'],
+                    k=stoch_params["k"],
+                    d=stoch_params["d"]
+                )
+                if stoch_result is not None and not stoch_result.empty:
+                    df_with_indicators = pd.concat([df_with_indicators, stoch_result], axis=1)
+            
+            # Calculate Bollinger Bands
+            bb_params = config["bollinger"]
+            if len(df_with_indicators) >= bb_params["n"]:
+                bb_result = ta.bbands(
+                    df_with_indicators['close'], 
+                    length=bb_params["n"], 
+                    std=bb_params["std"]
+                )
+                if bb_result is not None and not bb_result.empty:
+                    df_with_indicators = pd.concat([df_with_indicators, bb_result], axis=1)
+            
+            # Calculate ATR (Average True Range)
+            if len(df_with_indicators) >= config["atr"]:
+                df_with_indicators['atr'] = ta.atr(
+                    df_with_indicators['high'], 
+                    df_with_indicators['low'], 
+                    df_with_indicators['close'], 
+                    length=config["atr"]
+                )
+            
+            # Calculate CMF (Chaikin Money Flow)
+            if len(df_with_indicators) >= config["cmf"]:
+                df_with_indicators['cmf'] = ta.cmf(
+                    df_with_indicators['high'], 
+                    df_with_indicators['low'], 
+                    df_with_indicators['close'], 
+                    df_with_indicators['volume'], 
+                    length=config["cmf"]
+                )
+            
+            print(f"✅ Technical indicators added: {len(df_with_indicators.columns) - len(df.columns)} new columns")
+            return df_with_indicators
             
         except Exception as e:
-            print(f"❌ 필터링된 데이터 저장 실패: {e}")
-            return ""
+            print(f"❌ Error calculating technical indicators: {e}")
+            return df
     
-    def process_api_response(self, raw_data: Dict[str, Any], stock_code: str, chart_type: str, base_date: str = None, expected_start_date: str = None, expected_end_date: str = None) -> Dict[str, Any]:
+    def _save_filtered_data_csv(self, df: pd.DataFrame, stock_code: str, chart_type: str, 
+                               base_date: str = None, expected_start_date: str = None, 
+                               expected_end_date: str = None) -> str:
         """
-        키움 API 응답 처리 (새로운 로직)
+        Save filtered DataFrame to CSV format only
         
         Args:
-            raw_data: 키움 API 원본 응답
-            stock_code: 종목코드
-            chart_type: 차트 유형
-            base_date: 기준일자
-            expected_start_date: 예상 시작일 (YYYYMMDD)
-            expected_end_date: 예상 종료일 (YYYYMMDD)
+            df: Filtered DataFrame to save
+            stock_code: Stock symbol
+            chart_type: Chart type (minute, day, week, month, year)
+            base_date: Base date for the request
+            expected_start_date: Expected start date (YYYYMMDD)
+            expected_end_date: Expected end date (YYYYMMDD)
             
         Returns:
-            Dict: 처리 결과 (필터링된 데이터 또는 업그레이드 제안)
+            str: Saved CSV file path
         """
-        print(f"🔄 {chart_type.upper()} 차트 데이터 처리 시작: {stock_code}")
+        if df.empty:
+            print("⚠️  Empty DataFrame - not saving filtered data")
+            return ""
         
-        # 1. 원본 데이터 저장 (새로운 파일명 규칙)
-        raw_filepath = self.save_raw_data(raw_data, stock_code, chart_type, base_date)
+        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+        api_id = self.chart_configs.get(chart_type, {}).get("api_function", "unknown")
+        base_date_str = base_date if base_date else "nodate"
         
-        # 2. 기간 검증이 요청된 경우에만 처리
+        # Create filename with date range if available
         if expected_start_date and expected_end_date:
-            # 가장 오래된 날짜 찾기
-            oldest_date = self._find_oldest_date_in_raw_data(raw_data, chart_type)
+            date_range = f"{expected_start_date}_{expected_end_date}"
+            filename = f"{timestamp}_{api_id}_{stock_code}_{base_date_str}_{date_range}.csv"
+        else:
+            filename = f"{timestamp}_{api_id}_{stock_code}_{base_date_str}_all.csv"
+        
+        csv_filepath = self.filtered_dir / filename
+        
+        try:
+            # Save as CSV (human readable, Excel compatible)
+            df.to_csv(csv_filepath, index=False, encoding='utf-8-sig')
+            csv_size = csv_filepath.stat().st_size
             
-            if oldest_date:
-                print(f"📅 가장 오래된 데이터: {oldest_date}, 요구한 시작일: {expected_start_date}")
-                
-                # 가장 오래된 데이터가 요구한 시작일보다 미래인 경우 업그레이드 제안
-                if oldest_date > expected_start_date:
-                    upgrade_info = self._get_chart_upgrade_suggestion(chart_type)
-                    
-                    result = {
-                        "status": "upgrade_required",
-                        "message": f"초기 기간 데이터 부족: {oldest_date} > {expected_start_date}",
-                        "upgrade_suggestion": upgrade_info,
-                        "raw_file": raw_filepath
-                    }
-                    
-                    if upgrade_info["next_type"]:
-                        result["suggestion"] = f"get_{upgrade_info['next_type']}_chart(stock_code='{stock_code}', base_date='{expected_end_date}')"
-                    
-                    print(f"❌ 업그레이드 제안: {upgrade_info['description']}")
-                    return result
+            print(f"✅ CSV saved: {csv_filepath}")
+            print(f"📊 Saved {len(df)} records with {len(df.columns)} columns")
+            print(f"💾 CSV size: {csv_size:,} bytes")
             
-            # 3. 날짜 범위 필터링
-            filtered_records = self._filter_data_by_date_range(raw_data, chart_type, expected_start_date, expected_end_date)
+            return str(csv_filepath)
             
-            if filtered_records:
-                # 차트 유형에 맞게 날짜 형식 변환
-                converted_records = self._convert_date_format_for_chart_type(filtered_records, chart_type)
-                
-                # 변환된 데이터 저장
-                filtered_filepath = self.save_filtered_data(converted_records, stock_code, chart_type, base_date)
-                
-                # 추가 로직: 필터링된 레코드가 100개 초과시 업그레이드 제안
-                record_count = len(converted_records)
-                if record_count > 100:
-                    upgrade_info = self._get_chart_upgrade_suggestion(chart_type)
-                    
-                    print(f"⚠️  데이터 과다 ({record_count}개 > 100개): {upgrade_info['description']}")
-                    
-                    result = {
-                        "status": "upgrade_required",
-                        "message": f"데이터 과다 ({record_count}개 레코드 > 100개 권장): 더 큰 차트 간격 사용 권장",
-                        "upgrade_suggestion": upgrade_info,
-                        "data_count": record_count,
-                        "raw_file": raw_filepath,
-                        "filtered_file": filtered_filepath
-                    }
-                    
-                    if upgrade_info["next_type"]:
-                        result["suggestion"] = f"get_{upgrade_info['next_type']}_chart(stock_code='{stock_code}', base_date='{expected_end_date}', expected_start_date='{expected_start_date}', expected_end_date='{expected_end_date}')"
-                    
-                    return result
-                
-                # 100개 이하인 경우 정상 반환
-                return {
-                    "status": "success",
-                    "message": f"데이터 필터링 완료: {record_count}개 레코드",
-                    "data": converted_records,
-                    "raw_file": raw_filepath,
-                    "filtered_file": filtered_filepath
-                }
-            else:
-                return {
-                    "status": "no_data",
-                    "message": "지정된 날짜 범위에 데이터가 없습니다.",
-                    "raw_file": raw_filepath
-                }
-        
-        # 4. 기간 검증이 요청되지 않은 경우 원본 데이터 그대로 반환
-        return {
-            "status": "raw_data",
-            "message": "원본 데이터 저장 완료",
-            "data": raw_data,
-            "raw_file": raw_filepath
-        }
-    
-    def get_data_summary(self) -> Dict[str, Any]:
-        """현재 저장된 데이터 요약 정보 반환"""
-        summary = {
-            "raw_files": len(list(self.raw_dir.glob("*.json"))) if self.raw_dir.exists() else 0,
-            "filtered_files": len(list(self.filtered_dir.glob("*.json"))) if self.filtered_dir.exists() else 0,
-            "total_size_mb": 0,
-            "chart_types_supported": list(self.chart_configs.keys())
-        }
-        
-        # 총 크기 계산
-        total_size = 0
-        for dir_path in [self.raw_dir, self.filtered_dir]:
-            if dir_path.exists():
-                for file_path in dir_path.rglob("*.json"):
-                    total_size += file_path.stat().st_size
-        
-        summary["total_size_mb"] = round(total_size / (1024 * 1024), 2)
-        
-        return summary
+        except Exception as e:
+            print(f"❌ Failed to save CSV data: {e}")
+            return ""
 
 
-# 전역 데이터 매니저 인스턴스
+# Global data manager instance
 _data_manager = None
 
 def get_data_manager() -> StockDataManager:
-    """전역 데이터 매니저 인스턴스 반환"""
+    """Get global data manager instance"""
     global _data_manager
     if _data_manager is None:
         _data_manager = StockDataManager()
-    return _data_manager
-
-
-def process_api_response_for_tools(raw_data: Dict[str, Any], stock_code: str, chart_type: str, base_date: str = None, expected_start_date: str = None, expected_end_date: str = None) -> str:
-    """
-    키움 API 응답을 처리하고 결과를 반환 (LangChain 도구 전용)
-    
-    Args:
-        raw_data: 키움 API 원본 응답
-        stock_code: 종목코드
-        chart_type: 차트 유형
-        base_date: 기준일자
-        expected_start_date: 예상 시작일 (YYYYMMDD)
-        expected_end_date: 예상 종료일 (YYYYMMDD)
-        
-    Returns:
-        str: 처리 결과 JSON 문자열
-    """
-    try:
-        data_manager = get_data_manager()
-        result = data_manager.process_api_response(
-            raw_data, stock_code, chart_type, base_date, expected_start_date, expected_end_date
-        )
-        
-        return json.dumps(result, ensure_ascii=False)
-        
-    except Exception as e:
-        print(f"❌ 데이터 처리 오류: {e}")
-        # 오류 시 원본 데이터 반환
-        return json.dumps({
-            "status": "error",
-            "message": f"데이터 처리 오류: {str(e)}",
-            "data": raw_data
-        }, ensure_ascii=False) 
+    return _data_manager
