+++ agents/stock_price_agent/utils.py
